refactor: use kebab-case for codes as they might be served via http
