+++ .storybook/react/manager.ts
@@ -1,6 +1,6 @@
 import { addons } from "@storybook/manager-api";
 
-import daikinTheme from "../daikinTheme";
+import daikinTheme from "../daikin-theme";
 
 addons.setConfig({
   theme: daikinTheme,
