+++ src/tests/visual/clip-for.ts
