+++ src/tests/visual/index.ts
@@ -1,3 +1,3 @@
-export * from "./clipFor";
+export * from "./clip-for";
 export * from "./config";
 export * from "./runner";
